crawl4ai/requirements.txt
crawl4ai>=0.6.3
pyyaml>=6.0 